Move files into their own editable constraint folder
