+++ frontend/src/component/feature/FeatureStrategy/FeatureStrategyConstraints/EditableConstraint/LegalValuesSelector.tsx
